requirements.txt
sentence-transformers
redis
pymupdf
faiss-cpu
google-generativeai
python-dotenv
langchain_text_splitters
nltk
rank-bm25
fastapi
uvicorn